master.cfg: Specify branch for gnulib git repository
The octave-stable builder seems to have problems updating the gnulib repository. The log shows that it tries to fetch from a "stable" branch (which doesn't exist in the upstream repository).
Maybe specifying the branch explicitly will help with this build step.
Side note: Why is this build step even necessary? Won't the bootstrap script do that automatically?