Continue using cache for re-compiled compilers
By default, ccache includes the modification time of the compiler in the hash (see https://ccache.dev/manual/4.0.html#_configuration_options ). That means that the cache cannot be re-used for the compilers that are (re-)built as part of MXE Octave.
Use the version information instead to identify the used compiler. Afaict, the version string doesn't contain time information and should stay stable as long as the compiler version or configuration doesn't change.