master: store builds under OCTAVE_BUILD_ID
This OCTAVE_BUILD_ID is effectively the "buildnumber" of a builder
(e.g. "octave", the stable branch).  This number is not as unique as
"buildid", see

https://docs.buildbot.net/3.1.1/manual/configuration/properties.html#common-build-properties

The OCTAVE_HG_ID does not qualify anymore, as for mxe changes the very
same Octave hg-id is build again for reasons of convenience.