Pull from Octave mirror on GitHub. (#38)
The anonymous access over HTTPS to the Mercurial repository on Savannah is
getting more and more unreliable. (That is possibly because crawlers for LLMs
are hammering that server with a torrent of requests.)

Attempt to pull the repository from the mirror of the Octave repository on
GitHub instead.
That way, we only rely on the synchronization from the Mercurial repository on
Savannah to the GitHub mirror. But we don't rely on the Mercurial repository on
Savannah to be available each time a build starts or the buildbot scheduler
checks for updates.

The MXE Octave repository (on the Digital Ocean server) is still responsive
most of the time. (Maybe, the crawlers didn't find it yet.) And we don't have
a mirror for that (yet?). So, keep pulling the Mercurial repository for it.